Update to my current status
Adding in a huge commit here, which adds all of the work ive done up until now.

I'm working on ironing out some bugs but 90% of the code is here and ready.

currently won't write a file as some bytes are WAY to big. time to figure it out ig.